exception handling

+++ cleanFilenames.py
@@ -1,7 +1,8 @@
 import os
 import re
 
-os.remove("archive_filelist.csv")
+if os.path.isfile("archive_filelist.csv"):
+    os.remove("archive_filelist.csv")
 
 for filename in os.listdir("."):
     os.rename(filename, re.sub('^[+[0-9]+]_', '', filename))
